MyAPI-main/src/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
distlib==0.3.8
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
filelock==3.13.4
gunicorn==22.0.0
idna==3.7
packaging==24.0
pillow==10.3.0
platformdirs==4.2.1
psycopg==3.1.19
psycopg2==2.9.9
requests==2.31.0
sqlparse==0.5.0
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.0
whitenoise==6.6.0
